Misc clippy fixes
* `clippy::bool_assert_comparison`
* `clippy::field_reassign_with_default`
* `clippy::legacy_numeric_constants`
* `clippy::manual_clamp`
* `clippy::map_flatten`
* `clippy::needless_borrows_for_generic_args`
* `clippy::needless_lifetimes`
* `clippy::redundant_field_names`
* `clippy::to_string_in_format_args`
* `clippy::unnecessary_cast`
* `clippy::useless_format`

+++ crates/c-api/lib.rs
@@ -155,7 +155,7 @@ pub extern "C" fn resvg_options_set_resources_dir(opt: *mut resvg_options, path:
 /// Default: 96
 #[no_mangle]
 pub extern "C" fn resvg_options_set_dpi(opt: *mut resvg_options, dpi: f32) {
-    cast_opt(opt).dpi = dpi as f32;
+    cast_opt(opt).dpi = dpi;
 }
 
 /// @brief Sets the default font family.
